combine rules
