+++ models/src/main/scala/traderx/morphir/rulesengine/models/TradeOrder.scala
@@ -12,6 +12,7 @@ object TradeOrder{
   final case class TradeOrder(
     id: morphir.sdk.String.String,
     state: traderx.morphir.rulesengine.models.TradeState.TradeState,
+    desired: traderx.morphir.rulesengine.models.TradeState.TradeState,
     security: morphir.sdk.String.String,
     quantity: morphir.sdk.Basics.Int,
     accountId: morphir.sdk.String.String,
